+++ test/private/jsdocCommentToMember.test.js
@@ -6,34 +6,51 @@ const snapshot = require('snapshot-assertion');
 const codeToJsdocComments = require('../../private/codeToJsdocComments');
 const jsdocCommentToMember = require('../../private/jsdocCommentToMember');
 
+const TEST_CODE_FILE_PATH = '/a.js';
+
 module.exports = (tests) => {
   tests.add('`jsdocCommentToMember` with a JSDoc syntax error.', () => {
-    const [jsdocComment] = codeToJsdocComments('/** @tag [name */');
+    const [jsdocComment] = codeToJsdocComments(
+      '/** @tag [name */',
+      TEST_CODE_FILE_PATH
+    );
     strictEqual(jsdocCommentToMember(jsdocComment), undefined);
   });
 
   tests.add('`jsdocCommentToMember` with no description, no tags.', () => {
-    const [jsdocComment] = codeToJsdocComments('/** */');
+    const [jsdocComment] = codeToJsdocComments('/** */', TEST_CODE_FILE_PATH);
     strictEqual(jsdocCommentToMember(jsdocComment), undefined);
   });
 
   tests.add('`jsdocCommentToMember` with description, no tags.', () => {
-    const [jsdocComment] = codeToJsdocComments('/** Description. */');
+    const [jsdocComment] = codeToJsdocComments(
+      '/** Description. */',
+      TEST_CODE_FILE_PATH
+    );
     strictEqual(jsdocCommentToMember(jsdocComment), undefined);
   });
 
   tests.add('`jsdocCommentToMember` with a missing kind.', () => {
-    const [jsdocComment] = codeToJsdocComments('/** @name A */');
+    const [jsdocComment] = codeToJsdocComments(
+      '/** @name A */',
+      TEST_CODE_FILE_PATH
+    );
     strictEqual(jsdocCommentToMember(jsdocComment), undefined);
   });
 
   tests.add('`jsdocCommentToMember` with a missing name.', () => {
-    const [jsdocComment] = codeToJsdocComments('/** @kind member */');
+    const [jsdocComment] = codeToJsdocComments(
+      '/** @kind member */',
+      TEST_CODE_FILE_PATH
+    );
     strictEqual(jsdocCommentToMember(jsdocComment), undefined);
   });
 
   tests.add('`jsdocCommentToMember` with tag ignore.', () => {
-    const [jsdocComment] = codeToJsdocComments('/** @ignore */');
+    const [jsdocComment] = codeToJsdocComments(
+      '/** @ignore */',
+      TEST_CODE_FILE_PATH
+    );
     strictEqual(jsdocCommentToMember(jsdocComment), undefined);
   });
 
@@ -48,7 +65,7 @@ const a = true;
 const b = true;
 
 // Code after…`;
-    const [jsdocComment] = codeToJsdocComments(code);
+    const [jsdocComment] = codeToJsdocComments(code, TEST_CODE_FILE_PATH);
 
     let caughtError;
 
@@ -70,11 +87,14 @@ const b = true;
   });
 
   tests.add('`jsdocCommentToMember` with description.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * Description.
  * @kind member
  * @name A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -85,11 +105,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag desc, missing description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @desc
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -102,11 +125,14 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag desc.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @desc Description.
  * @kind member
  * @name A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -117,12 +143,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag desc, overriding description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * Description A.
  * @desc Description B.
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -137,12 +166,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag desc, overriding tag desc.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @desc Description A.
  * @desc Description B.
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -157,12 +189,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag desc, overriding tag description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @description Description A.
  * @desc Description B.
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -177,11 +212,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag description, missing description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @description
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -194,11 +232,14 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag description.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @description Description.
  * @kind member
  * @name A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -212,12 +253,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag description, overriding description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * Description A.
  * @description Description B.
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -232,12 +276,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag description, overriding tag desc.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @desc Description A.
  * @description Description B.
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -252,12 +299,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag description, overriding tag description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @description Description A.
  * @description Description B.
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -270,11 +320,14 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag kind, missing name.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind member
  * @kind
  * @name A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -286,10 +339,13 @@ const b = true;
   });
 
   tests.add('`jsdocCommentToMember` with tag kind.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind member
  * @name A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -300,11 +356,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag kind, overriding tag kind.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @kind member
  * @name A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -319,10 +378,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag kind, overriding tag typedef.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @typedef A
  * @kind member
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -337,10 +399,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag kind, overriding tag callback.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @callback A
  * @kind member
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -353,11 +418,14 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag name, missing name.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind member
  * @name A
  * @name
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -369,10 +437,13 @@ const b = true;
   });
 
   tests.add('`jsdocCommentToMember` with tag name.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind member
  * @name A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -383,11 +454,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag name, overriding tag name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @name B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -402,10 +476,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag name, overriding tag typedef.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @typedef A
  * @name B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -420,10 +497,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag name, overriding tag callback.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @callback A
  * @name B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -436,11 +516,14 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag type, missing type.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind member
  * @name A
  * @type
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -452,11 +535,14 @@ const b = true;
   });
 
   tests.add('`jsdocCommentToMember` with tag type.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind member
  * @name A
  * @type {boolean}
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -467,12 +553,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag type, overriding tag type.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @type {string}
  * @type {boolean}
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -487,10 +576,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag type, overriding tag typedef.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @typedef {string} A
  * @type {boolean}
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -505,10 +597,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag type, overriding tag callback.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @callback A
  * @type {SpecialFunctionType}
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -523,11 +618,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag typedef, missing name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @typedef
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -540,9 +638,12 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag typedef, name.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @typedef A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -556,9 +657,12 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag typedef, type, name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @typedef {boolean} A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -573,10 +677,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag typedef, overriding tag kind.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @typedef A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -591,10 +698,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag typedef, overriding tag type.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @type {string}
  * @typedef {boolean} A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -609,10 +719,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag typedef, overriding tag name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @name A
  * @typedef B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -627,10 +740,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag typedef, overriding tag typedef.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @typedef {string} A
  * @typedef {boolean} B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -645,10 +761,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag typedef, overriding tag callback.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @callback A
  * @typedef {boolean} B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -663,11 +782,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag callback, missing name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @callback
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -680,9 +802,12 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag callback, name.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @callback A
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -696,10 +821,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag callback, overriding tag kind.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @callback A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -714,10 +842,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag callback, overriding tag type.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @type {boolean}
  * @callback A
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -732,10 +863,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag callback, overriding tag name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @name A
  * @typedef B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -750,10 +884,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag callback, overriding tag typedef.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @typedef {boolean} A
  * @callback B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -768,10 +905,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag callback, overriding tag callback.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @callback A
  * @callback B
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -786,10 +926,13 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with an event without an `event:` name prefix.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind event
  * @name A#b
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -804,13 +947,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, missing name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg
  * @argument
  * @param
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -825,13 +971,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg a
  * @argument b
  * @param c
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -846,13 +995,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, type, name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg {boolean} a
  * @argument {boolean} b
  * @param {boolean} c
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -867,13 +1019,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, type (optional), name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg {boolean} [a]
  * @argument {boolean} [b]
  * @param {boolean} [c]
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -888,13 +1043,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, type (optional, default), name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg {boolean} [a=true]
  * @argument {boolean} [b=true]
  * @param {boolean} [c=true]
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -909,13 +1067,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg a Parameter description A.
  * @argument b Parameter description A.
  * @param c Parameter description A.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -930,13 +1091,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, type, name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg {boolean} a Parameter description A.
  * @argument {boolean} b Parameter description A.
  * @param {boolean} c Parameter description A.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -951,13 +1115,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, type (optional), name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg {boolean} [a] Parameter description A.
  * @argument {boolean} [b] Parameter description A.
  * @param {boolean} [c] Parameter description A.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -972,13 +1139,16 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag param synonyms, type (optional, default), name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @arg {boolean} [a=true] Parameter description A.
  * @argument {boolean} [b=true] Parameter description A.
  * @param {boolean} [c=true] Parameter description A.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -993,12 +1163,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, missing name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop
  * @property
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1013,12 +1186,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop a
  * @property b
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1033,12 +1209,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, type, name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop {boolean} a
  * @property {boolean} b
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1053,12 +1232,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, type (optional), name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop {boolean} [a]
  * @property {boolean} [b]
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1073,12 +1255,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, type (optional, default), name.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop {boolean} [a=true]
  * @property {boolean} [b=true]
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1093,12 +1278,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop a Property description A.
  * @property b Property description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1113,12 +1301,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, type, name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop {boolean} a Property description A.
  * @property {boolean} b Property description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1133,12 +1324,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, type (optional), name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop {boolean} [a] Property description A.
  * @property {boolean} [b] Property description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1153,12 +1347,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag prop synonyms, type (optional, default), name, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @prop {boolean} [a=true] Property description A.
  * @property {boolean} [b=true] Property description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1173,11 +1370,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag return, missing type or description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @return
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1190,11 +1390,14 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag return, type.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind function
  * @name A
  * @return {boolean}
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1208,12 +1411,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag return, type, overriding tag returns.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns {string} Description.
  * @return {boolean}
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1228,11 +1434,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag return, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @return Description.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1247,12 +1456,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag return, description, overriding tag returns.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns {boolean} Description A.
  * @return Description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1267,11 +1479,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag return, type, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @return {boolean} Description.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1286,12 +1501,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag return, type, description, overriding tag return.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @return {string} Description A.
  * @return {boolean} Description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1306,12 +1524,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag return, type, description, overriding tag returns.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns {string} Description A.
  * @return {boolean} Description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1326,11 +1547,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, missing type or description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1343,11 +1567,14 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag returns, type.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind function
  * @name A
  * @returns {boolean}
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1361,12 +1588,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, type, overriding tag return.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @return {string} Description.
  * @returns {boolean}
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1381,12 +1611,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, type, overriding tag returns.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns {string} Description.
  * @returns {boolean}
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1401,11 +1634,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns Description.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1420,12 +1656,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, description, overriding tag return.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @return {boolean} Description A.
  * @returns Description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1440,11 +1679,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, type, description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns {boolean} Description.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1459,12 +1701,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, type, description, overriding tag return.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @return {string} Description A.
  * @returns {boolean} Description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1479,12 +1724,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag returns, type, description, overriding tag returns.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind function
  * @name A
  * @returns {string} Description A.
  * @returns {boolean} Description B.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1499,12 +1747,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag fires synonyms, missing names.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind class
  * @name A
  * @emits
  * @fires
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1517,12 +1768,15 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag fires synonyms.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind class
  * @name A
  * @emits A#event:a
  * @fires A#event:b
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1536,14 +1790,17 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag fires synonyms, duplicate names.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind class
  * @name A
  * @emits A#event:a
  * @emits A#event:a
  * @fires A#event:b
  * @fires A#event:b
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1558,11 +1815,14 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag see, missing description.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @see
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1575,12 +1835,15 @@ const b = true;
   );
 
   tests.add('`jsdocCommentToMember` with tag see.', async () => {
-    const [jsdocComment] = codeToJsdocComments(`/**
+    const [jsdocComment] = codeToJsdocComments(
+      `/**
  * @kind member
  * @name A
  * @see See description A.
  * @see See description B.
- */`);
+ */`,
+      TEST_CODE_FILE_PATH
+    );
 
     await snapshot(
       JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1591,12 +1854,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag example, no caption, no content.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @example
  * @example
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1611,12 +1877,15 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag example, caption, no content.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @example <caption>Example A caption.</caption>
  * @example <caption>Example B caption.</caption>
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),
@@ -1631,14 +1900,17 @@ const b = true;
   tests.add(
     '`jsdocCommentToMember` with tag example, caption, content.',
     async () => {
-      const [jsdocComment] = codeToJsdocComments(`/**
+      const [jsdocComment] = codeToJsdocComments(
+        `/**
  * @kind member
  * @name A
  * @example <caption>Example A caption.</caption>
  * Example A content.
  * @example <caption>Example B caption.</caption>
  * Example B content.
- */`);
+ */`,
+        TEST_CODE_FILE_PATH
+      );
 
       await snapshot(
         JSON.stringify(jsdocCommentToMember(jsdocComment), null, 2),