+++ test/private/replaceJsdocLinks.test.js
@@ -16,7 +16,8 @@ module.exports = (tests) => {
 /**
  * @kind typedef
  * @name B
- */`
+ */`,
+    '/a.js'
   );
   const members = jsdocCommentsToMembers(jsdocComments);
   const outlinedMembers = outlineMembers(members);